Scene 8 changes to LD Footsteps and anims event
Added in 2 functions to be called for just Lao Da Anims so that his footsteps are louder added a anim event for his idle to stop footsteps.
Removed an extra audio source in a patrolling ai that was in its child objects.
Moved the Audio Listner to be where YY body is but in the rotation of the camera so the 3d spatial sound should come from where the thugs is to yy based on camera view, not model rotation.